requirements.txt
absl-py==0.12.0
argon==0.2.0
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
astunparse==1.6.3
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
boto3==1.17.27
botocore==1.20.27
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.6
decorator==4.4.2
defusedxml==0.7.1
dill==0.2.5
dj-database-url==0.5.0
Django @ file:///tmp/build/80754af9/django_1613762073373/work
django-allauth==0.44.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-estimators==0.2.1
django-heroku==0.3.1
django-imperavi==0.2.3
django-paging==0.2.5
django-rest-framework==0.1.0
django-sendgrid-v5==0.9.1
django-storages==1.11.1
django-summernote==0.8.11.6
django-templatetag-sugar==1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docutils==0.16
future==0.18.2
gast==0.4.0
google-pasta==0.2.0
grpcio==1.36.1
gunicorn==20.0.4
h5py==3.2.1
idna==2.10
importlib-metadata==3.7.2
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==0.10.0
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
ldif3==3.2.2
mysqlclient==2.0.3
numpy==1.20.1
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.2.3
parso==0.8.1
phonenumbers==8.12.19
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.17
protobuf==3.15.6
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
pyreadline==2.1
python-card-me==0.9.3
python-dateutil==2.8.1
python-decouple==3.4
python-http-client==3.3.2
python3-openid==3.2.0
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pywin32==300
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.4
sendgrid==6.6.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.7.0
soupsieve==2.2
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
starkbank-ecdsa==1.1.0
surlex==0.2.0
termcolor==1.1.0
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
unicodecsv==0.14.1
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.1
